+++ workflows_definition/RELEASES.md
@@ -251,3 +251,11 @@ Based on:
 - Speakeasy CLI 1.455.7 (2.480.1) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.13.0] workflows_definition
+
+## 2024-12-21 00:29:01
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/workflows-definition.yaml
+- Speakeasy CLI 1.459.2 (2.483.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.13.1] workflows_definition